backend/tensorflow_requirements.txt
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
six==1.11.0
Werkzeug==0.12.2
requests==2.18.4
pyAudioAnalysis==0.1.3
flake8==3.5.0
jupyter==1.0.0
python-magic==0.4.15
floyd-cli==0.10.28
Keras==2.1.2
pandas==0.21.1
numpy==1.13.3
tensorflow==1.4.1
h5py==2.7.1
seaborn==0.8.1
matplotlib==2.0.2
sounddevice==0.3.10
Pillow==4.2.1
